refactor: clean up IntegrationAPI
- remove not required class properties
- move API interface definitions to api_definitions.ts
- define a type for the WS client metadata object
